+++ index.ts
@@ -1,6 +1,7 @@
 import { appendOutput, setOutput } from "./output";
 import { execute } from "./tasks";
 export { ExecuteOptions } from "./tasks";
+export { registerActivities } from "./activities";
 
 export default {
   appendOutput,
